+++ src/main/java/ru/vyarus/guice/persist/orient/repository/command/ext/listen/ListenerTypeSupport.java
@@ -1,5 +1,6 @@
 package ru.vyarus.guice.persist.orient.repository.command.ext.listen;
 
+import com.google.inject.Injector;
 import com.orientechnologies.orient.core.command.OCommandRequest;
 import com.orientechnologies.orient.core.command.OCommandResultListener;
 import ru.vyarus.guice.persist.orient.repository.core.spi.parameter.ParamInfo;
@@ -7,11 +8,10 @@ import ru.vyarus.guice.persist.orient.repository.core.spi.parameter.ParamInfo;
 /**
  * Listener type handler.
  *
- * @param <T> listener type
  * @author Vyacheslav Rusakov
  * @since 29.09.2017
  */
-public interface ListenerTypeSupport<T> {
+public interface ListenerTypeSupport {
 
     /**
      * Check listener parameter correctness.
@@ -25,9 +25,14 @@ public interface ListenerTypeSupport<T> {
     /**
      * Checks listener compatibility with command object and wraps listener if required.
      *
-     * @param command  command object
-     * @param listener listener instance (passed in annotated parameter)
+     * @param command       command object
+     * @param listener      listener instance (passed in annotated parameter)
+     * @param injector      injector instance
+     * @param transactional true if listener execution must be wrapped with transaction
      * @return processed listener to apply to command
      */
-    OCommandResultListener processListener(OCommandRequest command, T listener);
+    OCommandResultListener processListener(OCommandRequest command,
+                                           Object listener,
+                                           Injector injector,
+                                           boolean transactional);
 }