+++ src/main/java/ru/vyarus/guice/persist/orient/repository/core/ext/service/result/converter/RecordConverter.java
@@ -1,5 +1,6 @@
 package ru.vyarus.guice.persist.orient.repository.core.ext.service.result.converter;
 
+import com.google.common.primitives.Primitives;
 import com.google.inject.Injector;
 import com.orientechnologies.orient.core.record.ORecord;
 import com.orientechnologies.orient.core.record.impl.ODocument;
@@ -10,6 +11,7 @@ import com.tinkerpop.blueprints.impls.orient.OrientBaseGraph;
 import ru.vyarus.guice.persist.orient.db.DatabaseManager;
 import ru.vyarus.guice.persist.orient.db.DbType;
 import ru.vyarus.guice.persist.orient.db.transaction.TransactionManager;
+import ru.vyarus.guice.persist.orient.repository.core.ext.util.ResultUtils;
 import ru.vyarus.guice.persist.orient.repository.core.result.ResultDescriptor;
 import ru.vyarus.guice.persist.orient.repository.core.result.ResultType;
 
@@ -18,7 +20,7 @@ import javax.inject.Singleton;
 import java.util.Set;
 
 /**
- * Used to convert single result (1 entity) using default result converter mechanism.
+ * Used to convert single record (1 entity) using default result converter mechanism.
  * This is required to apply conversion for orient listeners (async query or live query).
  * Service may be used directly (by injection).
  * <p>
@@ -26,27 +28,25 @@ import java.util.Set;
  * into object model class or vertex/edge (only if there are underlying transaction in thread(!) to avoid additional
  * connections opening). Default converter is applied only if object or graph type conversion is not required
  * (then default converter will check if projection to simple value is required).
- * <p>
- * Performs type conversion only from ODocument! It is not intended to be a universal converter for all types, only
- * for document (raw type). For example, you can't expect it to convert Vertex to ODocument or Model to Vertex.
  *
  * @author Vyacheslav Rusakov
+ * @see ResultConverter
  * @since 03.10.2017
  */
 @Singleton
-public class PlainResultConverter {
+public class RecordConverter {
 
-    private final ResultConverter defaultConverter;
+    private final ResultConverter defaultResultConverter;
     private final DatabaseManager databaseManager;
     private final TransactionManager transactionManager;
     private final Injector injector;
 
     @Inject
-    public PlainResultConverter(final ResultConverter defaultConverter,
-                                final DatabaseManager databaseManager,
-                                final TransactionManager transactionManager,
-                                final Injector injector) {
-        this.defaultConverter = defaultConverter;
+    public RecordConverter(final ResultConverter defaultResultConverter,
+                           final DatabaseManager databaseManager,
+                           final TransactionManager transactionManager,
+                           final Injector injector) {
+        this.defaultResultConverter = defaultResultConverter;
         this.databaseManager = databaseManager;
         this.transactionManager = transactionManager;
         this.injector = injector;
@@ -61,30 +61,31 @@ public class PlainResultConverter {
      * orient connection will perform automatic type conversion nad then custom result conversion is applied.
      * This service emulates the first part and calls result converter (so overall behaviour is near the same).
      *
-     * @param result     result object (most likely document)
+     * @param result     orient record
      * @param targetType target conversion type
      * @param <T>        target conversion type
      * @return converted object
      * @throws ResultConversionException if conversion is impossible
      */
-    public <T> T convert(final Object result, final Class<T> targetType) {
-        return convert(result, targetType, defaultConverter);
+    public <T> T convert(final ORecord result, final Class<T> targetType) {
+        return convert(result, targetType, defaultResultConverter);
     }
 
     /**
-     *
-     * @param result result object (most likely document)
+     * @param result     orient record
      * @param targetType target conversion type
-     * @param converter converter object
-     * @param <T> target conversion type
+     * @param converter  converter object
+     * @param <T>        target conversion type
      * @return converted object
      * @throws ResultConversionException if conversion is impossible
      */
     @SuppressWarnings("unchecked")
-    public <T> T convert(final Object result, final Class<T> targetType, final ResultConverter converter) {
+    public <T> T convert(final ORecord result, final Class<T> targetType, final ResultConverter converter) {
+        // wrap primitive, because result will always be object
+        final Class<T> target = Primitives.wrap(targetType);
         // do manual conversion to other types if required (usually this will be done automatically by connection
         // object, but async and live callbacks will always return documents)
-        T res = tryConversion(result, targetType);
+        T res = tryConversion(result, target);
         if (res != null) {
             return res;
         }
@@ -92,16 +93,17 @@ public class PlainResultConverter {
         // use converter
         final ResultDescriptor desc = new ResultDescriptor();
         // support void case for more universal usage
-        desc.returnType = targetType.equals(Void.class) || targetType.equals(void.class)
-                ? ResultType.VOID : ResultType.PLAIN;
-        desc.entityType = targetType;
-        desc.expectType = targetType;
-        return (T) converter.convert(desc, result);
+        desc.returnType = target.equals(Void.class) ? ResultType.VOID : ResultType.PLAIN;
+        desc.entityType = target;
+        desc.expectType = target;
+        res = (T) converter.convert(desc, result);
+        ResultUtils.check(res, target);
+        return res;
     }
 
-    private <T> T tryConversion(final Object result, final Class<T> targetType) {
+    private <T> T tryConversion(final ORecord result, final Class<T> targetType) {
         if (result == null
-                // no need for conversion of raw types
+                // no need for conversion of raw types (e.g. Document)
                 || ORecord.class.isAssignableFrom(targetType)
                 // without ongoing transaction we will have to open new connections for conversion (waste of resources)
                 || !transactionManager.isTransactionActive()