+++ src/test/groovy/ru/vyarus/guice/persist/orient/repository/command/live/advanced/AdvancedLiveCases.groovy
@@ -1,0 +1,34 @@
+package ru.vyarus.guice.persist.orient.repository.command.live.advanced
+
+import com.google.inject.ProvidedBy
+import com.google.inject.internal.DynamicSingletonProvider
+import com.google.inject.persist.Transactional
+import ru.vyarus.guice.persist.orient.repository.command.ext.listen.Listen
+import ru.vyarus.guice.persist.orient.repository.command.live.LiveQuery
+import ru.vyarus.guice.persist.orient.repository.command.live.mapper.LiveResultListener
+import ru.vyarus.guice.persist.orient.support.model.Model
+import ru.vyarus.guice.persist.orient.support.repository.mixin.crud.ObjectCrud
+import sun.security.provider.certpath.Vertex
+
+/**
+ * @author Vyacheslav Rusakov
+ * @since 10.10.2017
+ */
+@Transactional
+@ProvidedBy(DynamicSingletonProvider)
+interface AdvancedLiveCases extends ObjectCrud<Model> {
+
+    @LiveQuery("select from Model")
+    int subscribe(@Listen LiveResultListener<Model> listener)
+
+    @LiveQuery("select from Model where cnt > 1")
+    int subscribeConditional(@Listen LiveResultListener<Model> listener)
+
+    @LiveQuery("select from VertexModel")
+    int subscribeVertex(@Listen LiveResultListener<Vertex> listener)
+
+    // error - conversion will not be possible without enabled transaction (even if listener itself will
+    // be @Transactional guice bean
+    @LiveQuery("select from Model")
+    int subscribeNoTx(@Listen(transactional = false) LiveResultListener<Model> listener)
+}