+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 * Update to orient 2.2
 * Support [custom types installation](http://orientdb.com/docs/2.2/Object-2-Record-Java-Binding.html#custom-types)
   with new method: OrientModule#withCustomTypes(OObjectSerializer...) (#14)
+* (breaking) @Listen parameter now wraps provided listener with transaction by default (may be disabled: transactional = false)
 * Add @LiveQuery repository extension (live queries support)
+    - Listener parameter recognized by @Listen annotation
+    - custom LiveResultListener may be used to automatically convert record to object or vertex
+* New service RecordConverter may be used directly to convert ODocument to object or vertex and apply projection like in repositories methods
+(when document contains just one value it can be converted to value directly).    
+    
 
 ### 3.2.0 (2016-09-25)
 * Update to orient 2.1